notification once closed never open again for new message fix

+++ functions.php
@@ -1193,7 +1193,6 @@ function fep_add_message_participants( $message_id, $message, $inserted_message 
 				if( $participant != $inserted_message->post_author ){
 					fep_undelete_message( $inserted_message->post_parent, $participant);
 					delete_post_meta( $inserted_message->post_parent, '_fep_parent_read_by_'. $participant );
-					delete_user_option( $participant, '_fep_user_message_count' );
 				}
 			}
 		}		
@@ -1205,19 +1204,24 @@ function fep_add_message_participants( $message_id, $message, $inserted_message 
 					
 					if( ! in_array( $participant, fep_get_participants( $message_id ) )){
 						add_post_meta( $message_id, '_fep_participants', $participant );
-						delete_user_option( $participant, '_fep_user_message_count' );
+						if( 'publish' == $inserted_message->post_status ){
+							delete_user_option( $participant, '_fep_user_message_count' );
+							delete_user_option( $participant, '_fep_notification_dismiss' );
+						}
 					}
 				}
 			} else {
 				if( ! in_array( $message['message_to_id'], fep_get_participants( $message_id ) )){
 					add_post_meta( $message_id, '_fep_participants', $message['message_to_id'] );
-					delete_user_option( $message['message_to_id'], '_fep_user_message_count' );
+					if( 'publish' == $inserted_message->post_status ){
+						delete_user_option( $message['message_to_id'], '_fep_user_message_count' );
+						delete_user_option( $message['message_to_id'], '_fep_notification_dismiss' );
+					}
 				}
 			}
 		}		
 		if( ! in_array( $inserted_message->post_author, fep_get_participants( $message_id ) )){
 			add_post_meta( $message_id, '_fep_participants', $inserted_message->post_author );
-			delete_user_option( $inserted_message->post_author, '_fep_user_message_count' );
 		}
 			
 		fep_make_read( true, $message_id, $inserted_message->post_author );
@@ -1272,7 +1276,7 @@ function fep_send_message_transition_post_status( $new_status, $old_status, $pos
 		foreach( $participants as $participant ) 
 		{
 			delete_user_option( $participant, '_fep_user_message_count' );
-			if( $participant != $post->post_author ){
+			if( $participant != $post->post_author && 'publish' == $new_status ){
 				delete_user_option( $participant, '_fep_notification_dismiss' );
 			}
 		}
